+++ src/Phobos.COM.cpp
@@ -3,7 +3,7 @@
 #include <Helpers/Macro.h>
 
 #include <Locomotion/TestLocomotionClass.h>
-#include <Locomotion/SkilledLocomotionClass.h>
+#include <Locomotion/AdvancedDriveLocomotionClass.h>
 
 DEFINE_HOOK(0x6BD68D, WinMain_PhobosRegistrations, 0x6)
 {
@@ -13,7 +13,7 @@ DEFINE_HOOK(0x6BD68D, WinMain_PhobosRegistrations, 0x6)
 #ifdef CUSTOM_LOCO_EXAMPLE_ENABLED // Register the loco
 	RegisterFactoryForClass<TestLocomotionClass>();
 #endif
-	RegisterFactoryForClass<SkilledLocomotionClass>();
+	RegisterFactoryForClass<AdvancedDriveLocomotionClass>();
 
 	Debug::Log("COM registration done!\n");