+++ src/Utilities/TemplateDef.h
@@ -54,7 +54,7 @@
 #include <CRT.h>
 #include <LocomotionClass.h>
 #include <Locomotion/TestLocomotionClass.h>
-#include <Locomotion/SkilledLocomotionClass.h>
+#include <Locomotion/AdvancedDriveLocomotionClass.h>
 
 namespace detail
 {
@@ -1087,7 +1087,7 @@ if(_strcmpi(parser.value(), #name) == 0){ value = __uuidof(name ## LocomotionCla
 #ifdef CUSTOM_LOCO_EXAMPLE_ENABLED // Add semantic parsing for loco
 			PARSE_IF_IS_PHOBOS_LOCO(Test);
 #endif
-			PARSE_IF_IS_PHOBOS_LOCO(Skilled);
+			PARSE_IF_IS_PHOBOS_LOCO(AdvancedDrive);
 
 #undef PARSE_IF_IS_PHOBOS_LOCO